requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.9
opencv-python-headless>=4.8.0
anthropic>=0.34.0
Pillow>=10.0.0
numpy>=1.24.0
